API/add_db/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
arrow==1.2.1
async-timeout==4.0.1
attrs==21.2.0
awscli==1.20.56
beautifulsoup4==4.10.0
botocore==1.21.56
CacheControl==0.12.10
cachetools==4.2.4
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.3
cycler==0.11.0
docutils==0.15.2
elasticsearch==7.16.1
emoji==1.6.1
fake-useragent==0.1.11
firebase-admin==5.1.0
Flask==2.0.1
fonttools==4.28.2
frozenlist==1.2.0
fsspec==2021.11.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
googletransx==2.4.2
grpcio==1.42.0
grpcio-status==1.42.0
httplib2==0.20.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
msgpack==1.0.3
multidict==5.2.0
mysql-connector-python==8.0.27
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.4
python-socks==2.0.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
schedule==1.1.0
sendgrid==6.9.3
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
starkbank-ecdsa==2.0.3
stweet==2.0.0
tomli==1.2.2
tor-python-easy==0.1.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.1
yarl==1.7.2
